# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareBank Backend",
    "description": "Personalized Banking & Financial Wellness API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/planning/plans": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_planning_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FinancialPlanResponse"
                  },
                  "type": "array",
                  "title": "Response List Plans Api Planning Plans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_planning_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialPlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/planning/recurring-rules": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Create Recurring Rule",
        "operationId": "create_recurring_rule_api_planning_recurring_rules_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "List Recurring Rules",
        "operationId": "list_recurring_rules_api_planning_recurring_rules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecurringRuleResponse"
                  },
                  "title": "Response List Recurring Rules Api Planning Recurring Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/schedule-from-text": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Create Schedule From Text",
        "operationId": "create_schedule_from_text_api_planning_schedule_from_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleFromTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleFromTextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/planning/recurring-rules/{rule_id}": {
      "patch": {
        "tags": [
          "planning"
        ],
        "summary": "Update Recurring Rule",
        "operationId": "update_recurring_rule_api_planning_recurring_rules__rule_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/recurring-rules/{rule_id}/materialize": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Materialize Rule Once",
        "operationId": "materialize_rule_once_api_planning_recurring_rules__rule_id__materialize_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChecklistItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/scheduler/materialize-due": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Materialize Due Checklists",
        "operationId": "materialize_due_checklists_api_planning_scheduler_materialize_due_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulerMaterializeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/planning/checklist": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "List Checklist",
        "operationId": "list_checklist_api_planning_checklist_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChecklistItemResponse"
                  },
                  "title": "Response List Checklist Api Planning Checklist Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/checklist/{item_id}": {
      "patch": {
        "tags": [
          "planning"
        ],
        "summary": "Update Checklist Status",
        "operationId": "update_checklist_status_api_planning_checklist__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChecklistStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChecklistItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChecklistItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "recurring_rule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Rule Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "skipped",
              "failed"
            ],
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "recurring_rule_id",
          "title",
          "due_date",
          "amount",
          "status",
          "notes",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "ChecklistItemResponse"
      },
      "ChecklistStatusUpdateRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "skipped",
              "failed"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ChecklistStatusUpdateRequest"
      },
...
      "FinancialPlanCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "goal_type": {
            "type": "string",
            "enum": [
              "save",
              "invest",
              "debt",
              "expense_control",
              "custom"
            ],
            "title": "Goal Type",
            "default": "custom"
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount"
          },
          "monthly_budget": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Budget"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "FinancialPlanCreate"
      },
      "FinancialPlanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "goal_type": {
            "type": "string",
            "enum": [
              "save",
              "invest",
              "debt",
              "expense_control",
              "custom"
            ],
            "title": "Goal Type"
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount"
          },
          "monthly_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Budget"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "goal_type",
          "target_amount",
          "monthly_budget",
          "status",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "FinancialPlanResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecurringRuleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "rent",
              "bill",
              "gas",
              "saving",
              "investment",
              "custom"
            ],
            "title": "Category",
            "default": "custom"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "day_of_month": {
            "type": "integer",
            "maximum": 31.0,
            "minimum": 1.0,
            "title": "Day Of Month"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "autopay_enabled": {
            "type": "boolean",
            "title": "Autopay Enabled",
            "default": false
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": true
          },
          "reminder_days_before": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": 0.0,
            "title": "Reminder Days Before",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "title",
          "amount",
          "day_of_month"
        ],
        "title": "RecurringRuleCreate"
      },
      "RecurringRuleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "rent",
              "bill",
              "gas",
              "saving",
              "investment",
              "custom"
            ],
            "title": "Category"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "day_of_month": {
            "type": "integer",
            "title": "Day Of Month"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "next_run_date": {
            "type": "string",
            "format": "date",
            "title": "Next Run Date"
          },
          "autopay_enabled": {
            "type": "boolean",
            "title": "Autopay Enabled"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval"
          },
          "trusted_recurring": {
            "type": "boolean",
            "title": "Trusted Recurring"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "reminder_days_before": {
            "type": "integer",
            "title": "Reminder Days Before"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "plan_id",
          "title",
          "category",
          "amount",
          "day_of_month",
          "start_date",
          "next_run_date",
          "autopay_enabled",
          "requires_approval",
          "trusted_recurring",
          "is_active",
          "reminder_days_before",
          "created_at",
          "updated_at"
        ],
        "title": "RecurringRuleResponse"
      },
      "RecurringRuleUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rent",
                  "bill",
                  "gas",
                  "saving",
                  "investment",
                  "custom"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "day_of_month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Month"
          },
          "autopay_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autopay Enabled"
          },
          "requires_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires Approval"
          },
          "trusted_recurring": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trusted Recurring"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "reminder_days_before": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 14.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Days Before"
          }
        },
        "type": "object",
        "title": "RecurringRuleUpdate"
      },
...
      "ScheduleFromTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Text"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "default_amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Default Amount",
            "default": 0.0
          },
          "autopay_enabled": {
            "type": "boolean",
            "title": "Autopay Enabled",
            "default": false
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ScheduleFromTextRequest"
      },
      "ScheduleFromTextResponse": {
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/RecurringRuleResponse"
          },
          "extracted": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "extracted"
        ],
        "title": "ScheduleFromTextResponse"
      },
      "SchedulerMaterializeRequest": {
        "properties": {
          "until_days": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Until Days",
            "default": 14
          }
        },
        "type": "object",
        "title": "SchedulerMaterializeRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}